+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+*September 12, 2019*
+
+A quick bug fix on top of 0.0.2
+
+## v0.0.3
+
+### Bug fixes
+* [373](https://github.com/crypto-com/chain/pull/373) client-index auto-sync didn't work correctly
+
 *September 11, 2019*
 
 A small patches (mainly dependency bumps) to the released version
@@ -20,6 +29,13 @@ A small patches (mainly dependency bumps) to the released version
 
 The release is an incomplete alpha version meant to be deployed on the first iteration of the public testnet.
 There are no guarantees on future API and binary compatibility at this stage.
+
+## v0.1.0 (Unreleased)
+
+### Features
+
+### Bug Fixes
+* [365](https://github.com/crypto-com/chain/pull/365) Creating a transfer address also creates a staking address
 
 ## v0.0.1